House cleaning.
