+++ lms/djangoapps/linkedin/management/commands/login.py
@@ -3,7 +3,7 @@ Log into LinkedIn API.
 """
 from django.core.management.base import BaseCommand
 
-from .api import LinkedinAPI
+from . import LinkedinAPI
 
 
 class Command(BaseCommand):
